requirements-dev.txt
autoflake==1.3.1
black==19.3b0
codecov==2.0.15
coverage==4.5.1
dnspython==1.16.0
pycryptodome==3.8.1
pyOpenSSL==19.0.0
pytest==4.4.0
pytest-cov==2.6.1
Sphinx==2.0.1
sphinx-autodoc-typehints==1.6.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
